refactor(persona): the pull policy lives in persona/work_pull.rs (pure code motion)
WIP=1, the settle window, WIP=lanes (board-true), residency, the claim — one
policy, one 165-line file, out of the 5,300-line service loop it was edited in
four times today.
