+++ core/continuum-core/src/persona/mod.rs
@@ -99,6 +99,7 @@ pub mod self_task_generator;
 pub mod act_question;
 pub mod service_loop;
 pub mod wake_backlog;
+pub mod work_pull;
 pub mod staged_workspace;
 pub mod service_module;
 pub mod spawner;
